修复关服电网渲染报错
