+++ common/src/main/java/dev/dubhe/anvilcraft/client/renderer/PowerGridRenderer.java
@@ -2,7 +2,7 @@ package dev.dubhe.anvilcraft.client.renderer;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
-import dev.dubhe.anvilcraft.api.power.PowerGrid;
+import dev.dubhe.anvilcraft.api.power.SimplePowerGrid;
 import lombok.Getter;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -20,7 +20,7 @@ import java.util.Map;
 @Environment(EnvType.CLIENT)
 public class PowerGridRenderer {
     @Getter
-    private static Map<Integer, PowerGrid.SimplePowerGrid> grids = Collections.synchronizedMap(new HashMap<>());
+    private static Map<Integer, SimplePowerGrid> grids = Collections.synchronizedMap(new HashMap<>());
 
     /**
      * 渲染
@@ -29,7 +29,7 @@ public class PowerGridRenderer {
         if (Minecraft.getInstance().level == null) return;
         RandomSource random = Minecraft.getInstance().level.random;
         String level = Minecraft.getInstance().level.dimension().location().toString();
-        for (PowerGrid.SimplePowerGrid grid : PowerGridRenderer.grids.values()) {
+        for (SimplePowerGrid grid : PowerGridRenderer.grids.values()) {
             if (!grid.getLevel().equals(level)) continue;
             random.setSeed(grid.getHash());
             PowerGridRenderer.renderOutline(
@@ -40,7 +40,7 @@ public class PowerGridRenderer {
         }
     }
 
-    public static void cleanAllGrid() {
+    public static void clearAllGrid() {
         grids.clear();
     }
