+++ forge/src/main/java/dev/dubhe/anvilcraft/event/forge/ClientEventListener.java
@@ -1,7 +1,6 @@
 package dev.dubhe.anvilcraft.event.forge;
 
 
-import dev.dubhe.anvilcraft.client.renderer.PowerGridRenderer;
 import dev.dubhe.anvilcraft.util.IBlockHighlightUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
@@ -9,17 +8,15 @@ import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.client.event.RenderLevelStageEvent;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.level.LevelEvent;
 import org.jetbrains.annotations.NotNull;
 
-public class ClentEventListener {
+public class ClientEventListener {
     /**
      * 构造
      */
     @OnlyIn(Dist.CLIENT)
-    public ClentEventListener() {
+    public ClientEventListener() {
         MinecraftForge.EVENT_BUS.addListener(this::blockHighlight);
-        MinecraftForge.EVENT_BUS.addListener(this::onLevelUnload);
     }
 
     @OnlyIn(Dist.CLIENT)
@@ -30,10 +27,5 @@ public class ClentEventListener {
         if (level == null) return;
         IBlockHighlightUtil.render(level, Minecraft.getInstance().renderBuffers().bufferSource(),
             event.getPoseStack(), event.getCamera());
-    }
-
-    @OnlyIn(Dist.CLIENT)
-    private void onLevelUnload(@NotNull LevelEvent.Unload event) {
-        PowerGridRenderer.cleanAllGrid();
     }
 }
