+++ forge/src/main/java/dev/dubhe/anvilcraft/forge/AnvilCraftForge.java
@@ -2,7 +2,7 @@ package dev.dubhe.anvilcraft.forge;
 
 import dev.dubhe.anvilcraft.AnvilCraft;
 import dev.dubhe.anvilcraft.config.AnvilCraftConfig;
-import dev.dubhe.anvilcraft.event.forge.ClentEventListener;
+import dev.dubhe.anvilcraft.event.forge.ClientEventListener;
 import dev.dubhe.anvilcraft.init.ModCommands;
 import dev.dubhe.anvilcraft.init.forge.ModRecipeTypesForge;
 import dev.dubhe.anvilcraft.init.forge.ModVillagers;
@@ -30,7 +30,7 @@ public class AnvilCraftForge {
         bus.addListener(ModRecipeTypesForge::register);
         MinecraftForge.EVENT_BUS.addListener(AnvilCraftForge::registerCommand);
         try {
-            ClentEventListener ignore = new ClentEventListener();
+            ClientEventListener ignore = new ClientEventListener();
         } catch (NoSuchMethodError ignore) {
             AnvilCraft.LOGGER.debug("Server");
         }
